+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 ## [Unreleased]
 ### Added    
 * Experimental D-VARS sensitivity analysis
+* Sim.scalarOutVars() dict and Sim.noutvars
 ### Changed    
 * Removed mc_ and MC prefixes from all functions and classes
 ### Removed